backend/ml_webapp/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
drf_yasg==1.21.7
joblib==1.2.0
pandas
numpy
sklearn
django-cors-headers==3.14.0
